Fix warning in user controller and update user model
