+++ backend/database/databaseConnection.go
@@ -40,9 +40,9 @@ func DBinstance() *mongo.Client {
 var Client *mongo.Client = DBinstance()
 
 // OpenCollection is a  function makes a connection with a collection in the database
-func OpenCollection(client *mongo.Client, collectionName string) *mongo.Collection {
+func OpenCollection(client *mongo.Client, databaseName string, collectionName string) *mongo.Collection {
 
-	var collection *mongo.Collection = client.Database("cluster0").Collection(collectionName)
+	var collection *mongo.Collection = client.Database(databaseName).Collection(collectionName)
 
 	return collection
 }
