+++ backend/helpers/tokenHelper.go
@@ -1,126 +1,129 @@
 package helper
 
 import (
-    "context"
-    "fmt"
-    "log"
-    "os"
-    "time"
+	"context"
+	"fmt"
+	"log"
+	"os"
+	"time"
 
-    "backend/database"
+	"backend/database"
 
-    jwt "github.com/dgrijalva/jwt-go"
-    "go.mongodb.org/mongo-driver/bson"
-    "go.mongodb.org/mongo-driver/bson/primitive"
-    "go.mongodb.org/mongo-driver/mongo"
-    "go.mongodb.org/mongo-driver/mongo/options"
+	jwt "github.com/dgrijalva/jwt-go"
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/bson/primitive"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
 // SignedDetails
 type SignedDetails struct {
-    Email      string
-    First_name string
-    Last_name  string
-    Uid        string
-    jwt.StandardClaims
+	Email    string
+	Username string
+	Uid      string
+	jwt.StandardClaims
 }
 
-var userCollection *mongo.Collection = database.OpenCollection(database.Client, "user")
+var userCollection *mongo.Collection = database.OpenCollection(database.Client, "user_db", "user_accounts")
 
 var SECRET_KEY string = os.Getenv("SECRET_KEY")
 
 // GenerateAllTokens generates both teh detailed token and refresh token
-func GenerateAllTokens(email string, firstName string, lastName string, uid string) (signedToken string, signedRefreshToken string, err error) {
-    claims := &SignedDetails{
-        Email:      email,
-        First_name: firstName,
-        Last_name:  lastName,
-        Uid:        uid,
-        StandardClaims: jwt.StandardClaims{
-            ExpiresAt: time.Now().Local().Add(time.Hour * time.Duration(24)).Unix(),
-        },
-    }
+func GenerateAllTokens(email string, username string, uid string) (signedToken string, signedRefreshToken string, err error) {
+	claims := &SignedDetails{
+		Email:    email,
+		Username: username,
+		Uid:      uid,
+		StandardClaims: jwt.StandardClaims{
+			ExpiresAt: time.Now().Local().Add(time.Hour * time.Duration(24)).Unix(),
+		},
+	}
 
-    refreshClaims := &SignedDetails{
-        StandardClaims: jwt.StandardClaims{
-            ExpiresAt: time.Now().Local().Add(time.Hour * time.Duration(168)).Unix(),
-        },
-    }
+	refreshClaims := &SignedDetails{
+		StandardClaims: jwt.StandardClaims{
+			ExpiresAt: time.Now().Local().Add(time.Hour * time.Duration(168)).Unix(),
+		},
+	}
 
-    token, err := jwt.NewWithClaims(jwt.SigningMethodHS256, claims).SignedString([]byte(SECRET_KEY))
-    refreshToken, err := jwt.NewWithClaims(jwt.SigningMethodHS256, refreshClaims).SignedString([]byte(SECRET_KEY))
+	token, err := jwt.NewWithClaims(jwt.SigningMethodHS256, claims).SignedString([]byte(SECRET_KEY))
+	if err != nil {
+		log.Panic(err)
+		return
+	}
 
-    if err != nil {
-        log.Panic(err)
-        return
-    }
+	refreshToken, err := jwt.NewWithClaims(jwt.SigningMethodHS256, refreshClaims).SignedString([]byte(SECRET_KEY))
 
-    return token, refreshToken, err
+	if err != nil {
+		log.Panic(err)
+		return
+	}
+
+	return token, refreshToken, err
 }
 
-//ValidateToken validates the jwt token
+// ValidateToken validates the jwt token
 func ValidateToken(signedToken string) (claims *SignedDetails, msg string) {
-    token, err := jwt.ParseWithClaims(
-        signedToken,
-        &SignedDetails{},
-        func(token *jwt.Token) (interface{}, error) {
-            return []byte(SECRET_KEY), nil
-        },
-    )
+	token, err := jwt.ParseWithClaims(
+		signedToken,
+		&SignedDetails{},
+		func(token *jwt.Token) (interface{}, error) {
+			return []byte(SECRET_KEY), nil
+		},
+	)
 
-    if err != nil {
-        msg = err.Error()
-        return
-    }
+	if err != nil {
+		msg = err.Error()
+		return
+	}
 
-    claims, ok := token.Claims.(*SignedDetails)
-    if !ok {
-        msg = fmt.Sprintf("the token is invalid")
-        msg = err.Error()
-        return
-    }
+	claims, ok := token.Claims.(*SignedDetails)
+	if !ok {
+		msg = fmt.Sprintf("the token is invalid")
+		msg = err.Error()
+		return
+	}
 
-    if claims.ExpiresAt < time.Now().Local().Unix() {
-        msg = fmt.Sprintf("token is expired")
-        msg = err.Error()
-        return
-    }
+	if claims.ExpiresAt < time.Now().Local().Unix() {
+		msg = fmt.Sprintf("token is expired")
+		msg = err.Error()
+		return
+	}
 
-    return claims, msg
+	return claims, msg
 }
 
-//UpdateAllTokens renews the user tokens when they login
+// UpdateAllTokens renews the user tokens when they login
 func UpdateAllTokens(signedToken string, signedRefreshToken string, userId string) {
-    var ctx, cancel = context.WithTimeout(context.Background(), 100*time.Second)
+	var ctx, cancel = context.WithTimeout(context.Background(), 100*time.Second)
 
-    var updateObj primitive.D
+	var updateObj primitive.D
 
-    updateObj = append(updateObj, bson.E{"token", signedToken})
-    updateObj = append(updateObj, bson.E{"refresh_token", signedRefreshToken})
+	updateObj = append(updateObj, bson.E{"token", signedToken})
+	updateObj = append(updateObj, bson.E{"refresh_token", signedRefreshToken})
 
-    Updated_at, _ := time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))
-    updateObj = append(updateObj, bson.E{"updated_at", Updated_at})
+	Updated_at, _ := time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))
+	updateObj = append(updateObj, bson.E{"updated_at", Updated_at})
 
-    upsert := true
-    filter := bson.M{"user_id": userId}
-    opt := options.UpdateOptions{
-        Upsert: &upsert,
-    }
+	upsert := true
+	filter := bson.M{"user_id": userId}
+	opt := options.UpdateOptions{
+		Upsert: &upsert,
+	}
 
-    _, err := userCollection.UpdateOne(
-        ctx,
-        filter,
-        bson.D{
-            {"$set", updateObj},
-        },
-        &opt,
-    )
-    defer cancel()
+	_, err := userCollection.UpdateOne(
+		ctx,
+		filter,
+		bson.D{
+			{"$set", updateObj},
+		},
+		&opt,
+	)
+	defer cancel()
 
-    if err != nil {
-        log.Panic(err)
-        return
-    }
+	if err != nil {
+		log.Panic(err)
+		return
+	}
 
-    return
+	return
 }
