+++ backend/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"os"
 
-	middleware "backend/middleware"
+	// middleware "backend/middleware"
 	routes "backend/routes"
 
 	"github.com/gin-gonic/gin"
@@ -18,9 +18,9 @@ func main() {
 
 	router := gin.New()
 	router.Use(gin.Logger())
-	routes.UserRoutes(router)
+	routes.UserRoutes(router) // Creates User api routes
 
-	router.Use(middleware.Authentication())
+	// router.Use(middleware.Authentication())
 
 	// API-2
 	router.GET("/api-1", func(c *gin.Context) {
