+++ backend/middleware/authMiddleware.go
@@ -1,37 +1,37 @@
 package middleware
 
 import (
-    "fmt"
-    "net/http"
+	"fmt"
+	"net/http"
 
-    helper "backend/helpers"
+	helper "backend/helpers"
 
-    "github.com/gin-gonic/gin"
+	"github.com/gin-gonic/gin"
 )
 
 // Authz validates token and authorizes users
 func Authentication() gin.HandlerFunc {
-    return func(c *gin.Context) {
-        clientToken := c.Request.Header.Get("token")
-        if clientToken == "" {
-            c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("No Authorization header provided")})
-            c.Abort()
-            return
-        }
+	return func(c *gin.Context) {
+		fmt.Println("Authenticating")
+		clientToken := c.Request.Header.Get("token")
+		if clientToken == "" {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("No Authorization header provided")})
+			c.Abort()
+			return
+		}
 
-        claims, err := helper.ValidateToken(clientToken)
-        if err != "" {
-            c.JSON(http.StatusInternalServerError, gin.H{"error": err})
-            c.Abort()
-            return
-        }
+		claims, err := helper.ValidateToken(clientToken)
+		if err != "" {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": err})
+			c.Abort()
+			return
+		}
 
-        c.Set("email", claims.Email)
-        c.Set("first_name", claims.First_name)
-        c.Set("last_name", claims.Last_name)
-        c.Set("uid", claims.Uid)
+		c.Set("email", claims.Email)
+		c.Set("username", claims.Username)
+		c.Set("uid", claims.Uid)
 
-        c.Next()
+		c.Next()
 
-    }
+	}
 }
