+++ backend/models/userModel.go
@@ -1,22 +1,20 @@
 package models
 
 import (
-    "time"
+	"time"
 
-    "go.mongodb.org/mongo-driver/bson/primitive"
+	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
-//User is the model that governs all notes objects retrived or inserted into the DB
+// User is the model that governs all notes objects retrived or inserted into the DB
 type User struct {
-    ID            primitive.ObjectID `bson:"_id"`
-    First_name    *string            `json:"first_name" validate:"required,min=2,max=100"`
-    Last_name     *string            `json:"last_name" validate:"required,min=2,max=100"`
-    Password      *string            `json:"Password" validate:"required,min=6""`
-    Email         *string            `json:"email" validate:"email,required"`
-    Phone         *string            `json:"phone" validate:"required"`
-    Token         *string            `json:"token"`
-    Refresh_token *string            `json:"refresh_token"`
-    Created_at    time.Time          `json:"created_at"`
-    Updated_at    time.Time          `json:"updated_at"`
-    User_id       string             `json:"user_id"`
+	ID            primitive.ObjectID `bson:"_id"`
+	Username      string             `json:"username" validate:"required"`
+	Password      string             `json:"password" validate:"required,min=6"`
+	Email         string             `json:"email" validate:"email,required"`
+	Token         string             `json:"token"`
+	Refresh_token string             `json:"refresh_token"`
+	Created_at    time.Time          `json:"created_at"`
+	Updated_at    time.Time          `json:"updated_at"`
+	User_id       string             `json:"user_id"`
 }
