+++ backend/routes/userRoutes.go
@@ -1,13 +1,13 @@
 package routes
 
 import (
-    controller "backend/controllers"
+	controller "backend/controllers"
 
-    "github.com/gin-gonic/gin"
+	"github.com/gin-gonic/gin"
 )
 
-//UserRoutes function
+// UserRoutes function
 func UserRoutes(incomingRoutes *gin.Engine) {
-    incomingRoutes.POST("/users/signup", controller.SignUp())
-    incomingRoutes.POST("/users/login", controller.Login())
+	incomingRoutes.POST("/v1/signup", controller.SignUp())
+	// incomingRoutes.POST("/api-v1/login", controller.Login())
 }
